Added the spacings, adjusted the index and separated the css
Added spacings and a lot of variables that can be use to center the content inside the backbone if this one start and end not at the same place. It can has many other use cases. Removed the duplicate examples inside the index.html. Split the css into different files.